get all quotes working
